back/ZonaPetsBackend/apiLogin/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.37.0
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.2.3
